pyproject.toml
[tool.poetry]
name = "apple-health-parser"
version = "0.1.0"
description = "Python package to parse, analyse, and plot Apple HealthKit data"
authors = ["Alexandre Cirilo <38657258+alxdrcirilo@users.noreply.github.com>"]
repository = "https://github.com/alxdrcirilo/apple-health-parser"
license = "MIT"
readme = "README.md"
packages = [{ include = "apple_health_parser" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
kaleido = "0.2.1"
pandas = "^2.2.2"
plotly = "^5.22.0"
pydantic = "^2.7.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
pyclean = "^3.0.0"
mkdocstrings = { extras = ["python"], version = "^0.25.1" }
mkdocs-material = "^9.5.24"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
termynal = "^0.12.1"

[tool.poetry.scripts]
cli = "apple_health_parser.scripts.main:main"

[tool.semantic_release]
tag_format = "{version}"
version_toml = ["pyproject.toml:tool.poetry.version"]
major_on_zero = false
upload_to_PyPI = false
upload_to_release = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = ["I", "F401"]
ignore-init-module-imports = true

[tool.mypy]
ignore_missing_imports = true
